move World to wsv, fix roles feature, derive IntoSchema for CommittedBlock
