+++ core/src/tx.rs
@@ -2,7 +2,7 @@
 //!
 //! `Transaction` is the start of the Transaction lifecycle.
 
-use std::{cmp::min, collections::BTreeSet, time::Duration};
+use std::{cmp::min, time::Duration};
 
 use eyre::{Result, WrapErr};
 use iroha_crypto::{HashOf, SignaturesOf};
@@ -103,11 +103,6 @@ impl VersionedAcceptedTransaction {
         rejection_reason: TransactionRejectionReason,
     ) -> VersionedRejectedTransaction {
         self.into_inner_v1().reject(rejection_reason).into()
-    }
-
-    /// Checks if this transaction has already been committed or rejected.
-    pub fn is_in_blockchain<W: WorldTrait>(&self, wsv: &WorldStateView<W>) -> bool {
-        wsv.has_transaction(&self.hash())
     }
 
     /// # Errors
@@ -317,6 +312,22 @@ impl AcceptedTransaction {
     }
 }
 
+impl IsInBlockchain for VersionedAcceptedTransaction {
+    fn is_in_blockchain<W: WorldTrait>(&self, wsv: &WorldStateView<W>) -> bool {
+        wsv.has_transaction(&self.hash())
+    }
+}
+impl IsInBlockchain for VersionedValidTransaction {
+    fn is_in_blockchain<W: WorldTrait>(&self, wsv: &WorldStateView<W>) -> bool {
+        wsv.has_transaction(&self.hash())
+    }
+}
+impl IsInBlockchain for VersionedRejectedTransaction {
+    fn is_in_blockchain<W: WorldTrait>(&self, wsv: &WorldStateView<W>) -> bool {
+        wsv.has_transaction(&self.hash())
+    }
+}
+
 impl From<VersionedAcceptedTransaction> for VersionedTransaction {
     fn from(tx: VersionedAcceptedTransaction) -> Self {
         let tx: AcceptedTransaction = tx.into_inner_v1();
@@ -331,127 +342,6 @@ impl From<AcceptedTransaction> for Transaction {
             payload: transaction.payload,
             signatures: transaction.signatures.into_iter().collect(),
         }
-    }
-}
-
-impl From<VersionedValidTransaction> for VersionedTransaction {
-    fn from(transaction: VersionedValidTransaction) -> Self {
-        match transaction {
-            VersionedValidTransaction::V1(v1) => {
-                let transaction: ValidTransaction = v1.0;
-
-                let signatures = transaction
-                    .signatures
-                    .values()
-                    .iter()
-                    .cloned()
-                    .collect::<BTreeSet<_>>();
-                let tx = Transaction {
-                    payload: transaction.payload,
-                    signatures,
-                };
-                tx.into()
-            }
-        }
-    }
-}
-
-impl IsInBlockchain for VersionedRejectedTransaction {
-    fn is_in_blockchain<W: WorldTrait>(&self, wsv: &WorldStateView<W>) -> bool {
-        wsv.has_transaction(&self.hash())
-    }
-}
-
-declare_versioned_with_scale!(VersionedValidTransaction 1..2, Debug, Clone, iroha_derive::FromVariant);
-
-impl VersionedValidTransaction {
-    /// Same as [`as_v1`](`VersionedValidTransaction::as_v1()`) but also does conversion
-    pub const fn as_inner_v1(&self) -> &ValidTransaction {
-        match self {
-            Self::V1(v1) => &v1.0,
-        }
-    }
-
-    /// Same as [`as_inner_v1`](`VersionedValidTransaction::as_inner_v1()`) but returns mutable reference
-    pub fn as_mut_inner_v1(&mut self) -> &mut ValidTransaction {
-        match self {
-            Self::V1(v1) => &mut v1.0,
-        }
-    }
-
-    /// Same as [`into_v1`](`VersionedValidTransaction::into_v1()`) but also does conversion
-    pub fn into_inner_v1(self) -> ValidTransaction {
-        match self {
-            Self::V1(v1) => v1.0,
-        }
-    }
-
-    /// Apply instructions to the `WorldStateView<W>`.
-    /// # Errors
-    /// Fails if receives error during execution (should be fine after validation)
-    // XXX: Should it just return `()`?
-    pub fn proceed<W: WorldTrait>(&self, wsv: &WorldStateView<W>) -> Result<()> {
-        self.as_inner_v1().proceed(wsv)
-    }
-
-    /// Calculate transaction `Hash`.
-    pub fn hash(&self) -> HashOf<VersionedTransaction> {
-        self.as_inner_v1().hash().transmute()
-    }
-
-    /// Checks if this transaction has already been committed or rejected.
-    pub fn is_in_blockchain<W: WorldTrait>(&self, wsv: &WorldStateView<W>) -> bool {
-        wsv.has_transaction(&self.hash())
-    }
-
-    /// # Errors
-    /// Asserts specific instruction number of instruction in transaction constraint
-    pub fn check_instruction_len(&self, max_instruction_len: u64) -> Result<()> {
-        self.as_inner_v1()
-            .check_instruction_len(max_instruction_len)
-    }
-
-    /// Returns payload of transaction
-    pub const fn payload(&self) -> &Payload {
-        &self.as_inner_v1().payload
-    }
-}
-
-/// `ValidTransaction` represents trustfull Transaction state.
-#[version_with_scale(
-    n = 1,
-    versioned = "VersionedValidTransaction",
-    derive = "Debug, Clone"
-)]
-#[derive(Clone, Debug, Io, Encode, Decode)]
-pub struct ValidTransaction {
-    payload: Payload,
-    signatures: SignaturesOf<Payload>,
-}
-
-impl ValidTransaction {
-    /// # Errors
-    /// Asserts specific instruction number of instruction in transaction constraint
-    pub fn check_instruction_len(&self, max_instruction_len: u64) -> Result<()> {
-        self.payload.check_instruction_len(max_instruction_len)
-    }
-
-    /// Apply instructions to the `WorldStateView<W>`.
-    ///
-    /// # Errors
-    /// Can fail if execution of instructions fail
-    pub fn proceed<W: WorldTrait>(&self, wsv: &WorldStateView<W>) -> Result<()> {
-        for instruction in &self.payload.instructions {
-            instruction
-                .clone()
-                .execute(self.payload.account_id.clone(), wsv)?;
-        }
-        Ok(())
-    }
-
-    /// Calculate transaction `Hash`.
-    pub fn hash(&self) -> HashOf<Transaction> {
-        HashOf::new(&self.payload).transmute()
     }
 }
